requirements.txt
Django~=4.0.3
djangorestframework~=3.13.1
requests~=2.27.1
django-filter~=21.1
django-crispy-forms==1.14.0
